+++ lib/ellen/command_builder.rb
@@ -21,7 +21,7 @@ module Ruboty
     def options
       Slop.parse!(arguments, help: true) do
         on("dotenv", "Load .env before running.")
-        on("g", "generate", "Generate a new chatterbot with ./ellen/ directory if specified.")
+        on("g", "generate", "Generate a new chatterbot with ./ruboty/ directory if specified.")
         on("l", "load=", "Load a ruby file before running.")
       end.to_hash
     end