+++ spec/ellen/handlers/help_spec.rb
@@ -16,26 +16,26 @@ describe Ruboty::Handlers::Help do
 
     let(:body) do
       <<-EOS.strip_heredoc.strip
-        ellen /help( me)?\\z/i - Show this help message
-        ellen /ping\\z/i       - Return PONG to PING
-        ellen /who am i\\?/i   - Answer who you are
+        ruboty /help( me)?\\z/i - Show this help message
+        ruboty /ping\\z/i       - Return PONG to PING
+        ruboty /who am i\\?/i   - Answer who you are
       EOS
     end
 
-    it "responds to `@ellen help` and says each handler's usage" do
+    it "responds to `@ruboty help` and says each handler's usage" do
       robot.should_receive(:say).with(
         body: body,
         code: true,
         from: to,
         to: from,
         original: {
-          body: "@ellen help",
+          body: "@ruboty help",
           from: from,
           robot: robot,
           to: to,
         },
       )
-      robot.receive(body: "@ellen help", from: from, to: to)
+      robot.receive(body: "@ruboty help", from: from, to: to)
     end
   end
 end