+++ spec/spec_helper.rb
@@ -7,7 +7,7 @@ if ENV["CI"]
 end
 
 require "active_support/core_ext/string/strip"
-require "ellen"
+require "ruboty"
 
 RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true